backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
requests==2.31.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
ragas==0.1.9
deepeval==0.21.73
datasets==2.14.6
openai==1.3.7
langchain==0.1.0
langchain-openai==0.0.2
pandas==2.1.3
numpy==1.24.3
matplotlib==3.8.2
seaborn==0.13.0
python-multipart==0.0.6
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2